Vendor std::io::Cursor as buf::Cursor (closes #75)
Extracts the minimal parts of std::io::Cursor needed to support bytes use cases.

This is needed to support no_std as std::io is unavailable in these contexts.
It also (arguably) improves the API ergonomics by simplifying imports.

Per notes on #75, changes the u64 position to a u32 to save space.