"allocator" and "nightly" features (for no_std environments)
This commit gates all allocator-dependent features on an "allocator" feature.

It also adds a "nightly" feature which enables using the "allocator" feature in
no_std environments. This requires using #[feature(alloc)] which requires
nightly.

The "allocator" feature is automatically enabled when either the "std" or
"nightly" features are enabled.

Travis CI is configured to check that builds succeed with both the "nightly"
feature along with "std" and "nightly" in combination. To avoid the problem
of nightly changes breaking the build, these combinations are specifically
flagged as allowed failures in the Travis CI configuration.