impl `Zeroize` for `BytesMut` (optional)
`zeroize` is a popular Rust crate for securely zeroing memory:

https://crates.io/crates/zeroize

(I am the author)

Presently it's possible to use `Zeroize` with `BytesMut` via `DerefMut`
coercion (`Zeroize` is impl'd for `slice::IterMut`).

The implementation in this commit is more complete in that it handles
clearing data in the remaining capacity, in the event that it previously
contained secrets.

`zeroize` presently has a `bytes-preview` feature (which depends on
`bytes` 4.0) which provides this implementation:

https://docs.rs/zeroize/1.0.0/zeroize/#bytes-preview-feature-zeroize-support-for-bytesmut

I'd like to either upstream that implementation, i.e. this commit, or
drop the `bytes-preview` feature in `zeroize` as otherwise it seems
deref coercion is sufficient to get a "good enough" implementation
(and anyone who wants something better can make a newtype).