Implement BufQueue type
`BufQueue<B: Buf>` is a `VecDeque<B>` which implements `Buf` and
caches the remaining size.

This utility can be used to implement outgoing network buffer.
Imagine this use case:

```
type Header = Cursor<[u8; 5]>;
type Payload = Bytes;
type MyMessage = Chain<Header, Payload>;

outgoing_queue: BufDeque<MyMessage>;
```

New messages can be queues by simply appending to this `BufQueue`,
and the queue can be flushed by invoking `bytes_vectored` on the
queue object.

And both enqueue and flush operations are zero-copy.

`remaining` cache is needed to optimize `bytes` operation, which
might be expensive if the queue is long.

I'm not 100% sure this is how `Bytes` is meant to be used, but this
`BufQueue` type seems to be the most natural way.