WIP: Changes BytesMut not to rely on Vec + Arc
This changes the internal BytesMut representation to allocate memory on
its own, and not to use Arc + Vec under the hood.

Doing this allows to have a single representation which is shareable (by
being able to modify an embedded refcount) with zero-cost right from
the start. Not representation changes have to be performed over the
lifetime of `BytesMut`. Therefore `BytesMut::freeze()` is now also free.

The change also opens the door for adding support for buffer pools or
custom allocators, since an allocator is directly called. Adding this
support would mostly mean having to pass an allocator to BytesMut
which is used called instead of a fixed function. This would however be
a different change.

The change is WIP, and misses the following parts:
- The buffer growth strategy is not exactly the same as in the old
  version. It needs to be reviewed what the best strategy is.
- Some situations where an empty buffer (which has no backing
  storage) is manipulated might have bugs. This needs another
  round of review, and some additional tests.