Avoid extraneous call to remaining in get_u8/i8
I have a chained datastructure that can dereference into a slice quite fast, but remaining requires walking the chain, which is relatively slow.

I would hazard a guess that this is the more common perf characteristic of non-contiguous impls of Buf. 

Separately, this brings these impls in line with all the other get_ impls, which have a fast path with [u8].get(...), which does the same bounds check that I do in this patch

I think clippy will say to use non_empty, here, I think in this case this is more clear, but I will do whatever is preferred.