impl From<Box<[u8]>> for Bytes
This missing impl was an inconvenience to me, and there seemed to be
no reason for it not to exist.

The From<Vec<>> impl in fact turned the Vec into a Box almost
immediately.  I moved the check for empty slices into the new Box impl
since it is needed there too, and added a test case.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>