Copy only unread bytes in BytesMut::reserve
When reserving extra space in the Vec representation, there may not be
enough space to efficiently reuse the buffer. Currently, the entire
buffer is copied, including bytes that have already been read. Instead,
this change copies only the unread bytes to the beginning of the new
buffer. As a result, fewer bytes are copied and there is more unused
capacity.

It's important to note that the size of the new buffer will still be
double the size of the current buffer, including bytes already read.
That has not changed.

Before this change:

```
            read       unused capacity
             │           │
            ─┴─        ──┴──
old buffer: ░░░████████░░░░░

new buffer: ░░░████████░░░░░░░░░░░░░░░░░░░░░
            ─┬─        ──┬──────────────────
             │           │
            read       unused capacity
```

After this change:

```
            read       unused capacity
             │           │
            ─┴─        ──┴──
old buffer: ░░░████████░░░░░

new buffer: ████████░░░░░░░░░░░░░░░░░░░░░░░░
                    ──┬─────────────────────
                      │
                    unused capacity
```